Avoid accessing the enemyRobots vector out of bounds
Defenders are assigned to an enemy robot using the enemyRobot vector, but it was not checked whether there were sufficient enemies to give both defenders an enemy. This caused the enemyRobot to be set to an optional containing a null pointer instead of an empty optional vector. Later, in getRobotScoreForDistance, in Dealer.cpp, the distance to this enemy robot is calculated, which eventually fails as calculating the distance between a vector2 and NULL is not possible.