Use PIDType to increase velocity when keeper blocks ball
When the ball is shot towards the goal, the keepers PIDType is set to keeper. In GoToPos, it is checked whether the robots PIDTYpe is keeper, if so, and the robot is far enough away from its target position, its velocity will be set to max to more quickly intercept the ball.