Change Keeper position in KeeperBlockBall
When going to default position, stand 0.2 in front of goal instead of exactly on the line to prevent the robot from moving outside of the cameras vision. Also, change the location the keeper goes to when ball is neither shot nor near robot to be at the same y as the ball (clamped in the goal), instead of between the ball and the centre of the goal. This gives better positioning, as the keeper would otherwise would almost always stand nearly in the middle of the goal.