Use robot angle to determine keeper position instead of robot and ball position
Using the robot angle is more precise and consistent then using robot and ball location, as ball location can change quite a bit without the actual target position changing. Using robot angle prevents this.

Also, increased the length of the vectors to fieldlength instead of half of it to avoid missing collisions.