Replaced value by contant
When testing in real life, the robot drives out of the field and the AI loses the robot when you put it exactly on the goal line. Therefore a constant was added of .2 which makes to robot stand about its diameter more inside the field. This prevents the AI from losing the keeper on a real life field