Comment out TestPlay in application manager and note this in TestPlay
TestPlay should not be available as a play during regular play. To avoid having this happen, this play is commented out in the application manager and can not be selected. A comment in application manager and testPlay itself point to this to make sure people know what to do to enable the testplay for testing.