Use rotate skill when not at correct angle near ball in GetBall
Previously, when the robot was close to the ball but at the wrong angle, the target position would be set to the current position of the robot, trying to make it stand still while it finishes rotating. This is okay, but the robot does not quite stand still during this time, as neither getting the robots position nor moving to it are flawless, introducing some stuttering and making the rotation less clean. By instead using the rotate skill, it is guaranteed that the robot actually stands still and rotates smoothly towards the ball. As soon as the angle is correct, the tactic resets and the robot moves towards the ball again.