Do not touch ball before angle of robot is correct
If the ball is close to the robot but the robot is not facing towards the ball, it can happen that the robot hits the ball with its side in getBall, causing the ball to be pushed away. To avoid this, we ensure prevent the robot from getting too close to the ball while its angle is incorrect. As long as the angle is too far off, we set the robot to go to its own position, which makes the tactic go to the next skill: rotate (since goToPos is immediately successful). The robot then rotates until it is within a margin of HAS_BALL_ANGLE_ERROR_MARGIN * M_PI, when the tactic resets and the robot is told to go to the position of the ball. This ensures that the robot always faces the right direction before actually making contact with the ball.