Re-enable and fix limitAngularVel
When playing as right, the robot angle needs to be adjusted. This was not done in limitAngularVel yet, which caused the returned angle to be off by pi when playing as right. In turn, this lead to the AnglePID returning a very large output, as the difference between the current and target angle was very large due to the wrongly calculated angle from limitAngularVel. This caused the robot to move erratically and to rotate very quickly.