Add error margin to ball trajectory and robot aim calculations
In this tactic, it is checked whether 1) the ball is moving towards the goal or 2) an enemy is aiming towards the goal (while having a ball). The calculations for this are obviously never perfect, so it can happen that the keeper think the ball will just barely miss the goal even though it will actually go in. To avoid this, the goal is made artificially bigger for these calculations, to make the keeper intercept the ball even if it is expected to just go past the goal. The y-position for the keeper is then clamped between the actual goal (with some padding) to avoid the keeper standing (partly) outside the goal.