Make keeper look towards ball
By making the keeper look towards the ball the ball will hit the keeper on the front assembly, which reduces the chance of the ball reflecting in the goal, as the front assembly is straight as opposed to the curved sides of the robot. Furthermore, this provides more damping and increases the chance that the keeper is able to quickly grab and pass the ball.