Remove checks for side while creating field
When receiving world data and sending robot commands, everything needs to be rotated when playing as right, to ensure this information matches with the field (which is the same regardless of side). This can be unintuitive. In previous commits, this was attempted to be removed by actually changing the field when sides were switched (I.e. when playing as left, our goal was (-6,0) and (6,0) when playing as right). This removed the need to rotate incoming and outgoing information, simplifying part of the code. However, it introduced other difficulties. Mainly, calculations within STP would now need to consider side (or new functions would need to be written to automatically do this). For example, moving a keeper 20cm away from its goal used to be done by adding 0.2 to its x coordinate. When playing as right, however, this would move the keeper 20cm behind its goal when using relative coordinates. The same goes for, for example, constructing keeper arcs etc. In conclusion, it was deemed that the original approach of rotating incoming and outgoing information was better as it allows for easier programming everywhere in between.