Set keeperid in the interfaceGameState if the keeper role is distributed
During a game, one specific robot has to be the keeper (unless explicitly changed through the gamecontroller). Before, the default keeper Id was set to 0. Meaning that if robot 0 was in the field, it would get the keeper role. However, if this id was not in the field, another robot would get the keeper role, but which id this is would not be stored. However, it is important to know which id the keeper has to comply with certain rules (such as not being in the defense area unless the robot is keeper).

Now, the id of the robot that the keeper role is distributed to is saved in the interfacegamestate.keeperid, ensuring that we always know which robot is keeper.