Change default keeper id to -1 from 0
With a default keeper id of 0, robot 0 (if in play) will be given the keeper role, regardless of circumstances. This is problematic when testing as this means that robot 0 becomes keeper even if it is in a terrible position to do so.

When playing at the robocup, we need to declare which robot will be our keeper before the game starts. This is where the default keeper id is necessary and should be set to whichever id we want our keeper to be.