Split pointIsInDefenseArea to pointIsInOurDefenseArea and pointIsInTheirDefenseArea and add generic pointIsInDefenseArea
Previously, there was one function (pointIsInDefenseArea) that was used to check whether a point was in our, their, or either defense area by calling it with isOurDefenceArea with either true or false, and calling it twice when checking for either area. Although this works, it is error prone to have to supply the isOurDefenceArea bool, and it is cumbersome to call this function twice to check if a point is in either defense area (and one might forget to change isOurDefenceArea in the second call). Thus, this function is refactored in this PR to check whether the given point is in either defense area. The functions pointIsInOurDefenseArea and pointIsInTheirDefenseArea are added to check for our/their defense area in a more explicit manner.