Keep track of who is the ball placer in the gameState and set this in the dealer
For the pathing, it is important to know which robot is allowed to go where. This means we need to know who the keeper and who the ball placer are. This is stored in the game state. With this PR, the gamestate also tracks the ball placer in addition to the keeper. These ids are set in the dealer.