Update interfaceGameState when play is selected in interface
During a game, game states are updated as the autoref sends commands. These gamestates can be used to determine which rules currently apply. I.e. max robot speeds change between rule sets and a robot should be able to leave the field during ball placement. When using the interface to select plays, it is useful to keep track of which gamestate would currently be active during this play, to make sure the robots behave as they would during a real game. For example, during the stop game state, which is active during the AggressiveStopFormation play, robots have to avoid touching the ball. If the game state is not updated when a play is manually selected, there is no way of knowing that this rule apply.

In this PR, it is made sure that the relevant gamestate is active when a play is manually selected.