Clear stpInfos when (re-)distributing roles
When the roles need to be distributed (when a play is selected or when a robot is added/removed), it is important to clear the stpInfo's to make sure that the robots are no longer assigned to an stpInfo from before. Without clearing this, it can happen that if a robot is removed, it is still assigned in an stpInfo even though the robot is not on the field. This is problematic as it can cause issues when it is attempted to retrieve information from this robot.