Only update role if the role has a robot assigned to it
Without this check, it would be attempted to update the role, which would result in an error as it is checked there whether there is a robot assigned to this role whilst updating. By adding this check here, unnecessary function calls are avoided, and the role update will no longer give an error whenever there are less than 11 robots in play.