Add map to keep track of previous ball posession
Since the data we get wrt us having the ball is quite noisy, we can not just 1-on-1 use that to determine what our robots should do. This commit adds a functionality to instead keep track of a score that indicates how likely it is we have the ball. Each time we get a signal that indicates we have the ball (be it vision, ballsensor, dribblerencoder), we add to the score, and vice versa. We can then simply check if the score is high (= more signals that say we have the ball than not), which we interpret to mean that we indeed have the ball. This makes sure we don't think we have lost the ball just because vision thought we lost it for 1 tick due to noise etc.