Create function to project pos out of defense area on line
This function is useful when a point needs to be projected outside of the defense area, but you want to do so in a certain direction. For example, when standing in between a robot and the ball at a distance that would be in the defense area, you want to project that point outside of the defense area while still remaining on the line from the robot to the ball.