Add check to see if ball is too close to enemy to block normally
If the ball is too close (i.e. 0.5m away from the ball is super close/behind the enemy or 0.5m away from the enemy is behind the ball), place the robot at a distance of 4* the robot radius away from the enemy. This makes sure that the robot is still in between the ball and the robot whenever possible.

4 * the radius is slightly larger then technically necessary (2*radius is the actual minimum) but it is also important to make sure we do not crash  into the other robot. Hence, some extra padding was added.