Scale max velocity with distance to ball
The previous max velocity worked fine when close to the ball, but would be too slow if the robot is very far away from the ball. This normally is unlikely to happen, as orbit is used after getting the ball. However, if, for some reason, the ball would be further away while doing orbit, this makes sure that the robot goes a bit faster while far away. Furthermore, this allows orbit to be better used in the GetBehindBallInDirection tactic, where the robot is ofter further away when orbit starts.

This change does not really affect the behaviour of orbit in most situations, as the maxVel will still be 0.65 there (as this is the minimum).