Block robot handling situation happening near our defense zone
The robot positions itself between the enemy robot and a point that the robot has to defend at a given distance from the enemy robot (the distances can be CLOSE, HALFWAY, FAR).

If the point that the robot has to defend is inside the defense zone the location of the point is being projected to the border of the defense zone.

If the enemy robot is too close to our defense area so that our robot cannot position itself between the enemy robot and the defense zone then our robot will position itself in the proximity of the enemy robot.

The distance from the defense zone is calculated using getDistanceToDefenseZone from FieldComputations. There is also boolean checking whether the enemy robot is closer to our goal than our goal to our penalty line (isBelowPenaltyLine).