Check if role is actually initialised before trying to access it
Roles is an array of size 11, for 11 roles. Some plays now use 4 roles (for spain). Thus, there will be several nullptrs in this array that we need to check for.

Note: for flexibility, we could also use a vector instead of an array to avoid needing to check for nullptr. If we move back to all plays having 11 roles though, the 11-sized array is good.