Make scoreProfile argument const in position scoring
These profiles are defined as constexpr's. Passing them as a non-const reference generates an error when building, as these profiles can not be altered. The reason that this did not generate errors before, is that these functions were only called from another function, which was passed the profile by value (which creates a copy). Without this change, these functions could not be called directly using a profile as an argument, which can be useful.