Make attack valid play to pick/keep as long as they do not have ball
Even if the ball is not closest to us, it might be a good idea to attack (since the enemy might be slower/dumber/etc.). Deciding whether attacking if the ball is closer to them is actually a good idea, should be done using the score. This change ensures that attack is a valid play as long as they do not have the ball.