End attack play after attacker is finished.
Once the ball has been shot, attack should end, as the purpose of the play is fulfilled. Without this check, attack would keep running until they have the ball (and attack becomes invalid). Since the attacker's role is finished after it shoots, it will not do anything and just stand still, which is undesirable. With this change, the play will end after shooting as expected, and the next play can be executed.