Implement keeper priority flag to have the largest weight
This case was missing here, which caused warnings regarding unhandled dealerflags. Technically, it should not matter what this weight is, as long as it is non-zero, as the keeper is divided as the very first role anyways. Due to this, the keeper flags and the scores they lead to are never compared to other role scores, which means that multiplying each score by this weight does not change the outcome of the role assignments. However, this flag does need some value, and perhaps this number will be important in later implementations. Therefore, I chose to set this weight to 1000, as the keeper is generally regarded as the most important role, and should therefore have the highest priority weight.