Set refGameState to normal play after kickoff/throw-in
After the ref sends the kickoff/freekick command, the game continues as normal as soon as the ball has been brought back into play. At this point, we can behave normally again (except that the kicker is not yet allowed to touch the ball, but this is not taken into account for now as this is very unlikely to matter). To do so, we must change the gamestate to normal play (using the NORMAL_START command, in this case).

Note: technically, the ball is in play as soon as the ball has moved 5cm. However, this is more difficult to detect than it moving, so this is not done for now.