Improve position scoring and remove NFParam usage
This commit improves the way that positions are scored. This is mainly done by ensuring that the computations make sense, and that the different scores are balanced (i.e. tend to fall in the same range). For example, non-linear transformations are used to ensure that goalShotScores do not end up too low, as it is extremely unlikely that the goalShotScore will be nearly perfect, whereas the open or LoS score can quite easily be perfect. Based on testing, these computations and transformations appear to work quite well.