Remove rotate and always face towards ball in receive tactic
Since this is an end tactic, rotating can always be done during GoToPos. However, it is undesirable to do the rotate skill, as this prevents the robot from moving, which might be needed. Secondly, there is not really a point to not always looking towards the ball. It is safer to do so as this allows you to always be ready to receive the ball.