Create function projectPositionIntoFieldOnLine
Similar to projectPositionOutsidefDefenseArea, ths function projects a given point into the field, on a given line. In case the point is already in the field, it is simply returned. This commit also makes sure that projectPositionOutsidefDefenseArea simply returns the point if it is already outside the defense area.

To do this, the utility function lineIntersectionWithField was also created.