Implement basic passing computations
This function can calculate the best(-ish) pass location at the moment. It is still quite simple, and should be improved (TODO's have been added to make this clear).

The pass location is picked as follows:
- A grid of points is created as possible pass locations (for now, 9 points spread evenly on the field)
- For each point, it is checked whether it is possible to pass there based on several criteria
-- The position should be valid (in the field, not in defense area)
-- The position should be quicker to reach for the robot than for the ball (to make sure the pass can actually arrive)
-- The line of sight score to this position should be > 10 (essentially, there shouldnt be an enemy exactly on this path)
- For each point that fulfills these, the position is scored based on openness, line of sight, and goal shot score
- The point with the highest score is picked
If no pass can be found, the location of the robot deepest in the field is returned