Reduce no. roles to 4 and use new pass computations
In line with other plays, the no. roles is reduced to 4 in attackingPass. Furthermore, this play now uses the new pass computations to determine where to pass. This pass location is then used to set the position to move to for the receiver, and position to shoot at for the passer. The remaining robot is placed in the midfield on a safe position. The pass location is only calculated if there is not already one. This makes it easier for passes to be timed well, and allows for more accurate scoring of the play.

Note that this is a very first functional draft of implementing passing- and not perfect by any means. Further development is in progress, but this should be a decent starting point.