Score attacking pass based on scoring potential and risk
To score this play, we calculate the goalShotScore, which is then multiplied by the Line of Sight and Openness score (both normalised between 0 and 1). This tries to give the play an accurate score in comparison with attack, which is the main alternative. AttackingPass will only be preferred over attack if the goalShotScore of the calculated pass location is greater than the goalShotScore of the current (ball) position, after adjusting for the risk of the pass.