Use new GoalShot profile while scoring Attack
This is the same profile using in scoring attackingPass, which only looks at the goalshot, which is the most relevant metric for determining how likely it is to score a goal from a certain point. By using the same profile, we make sure that the scores are comparable between these two plays.