Set min_kick_power to 2
If kick_power is too low, the kicker literally does not hit the ball. Since we do not want this to happen, nor do we want to barely kick the ball, the min_kick_force is increased to 2. The exact value of this could be tuned better, but this suffices.