Simplify determineKickForce and reduce pass factor
Passing was too fast on the field, which caused the ball to bounce off or miss the robot. This commit reduces the kickForce factor, which will reduce the pass speed.

Note: this passForce could maybe be higher for the sim, as robots are more accurate there, but this would be a minor difference and would need more testing.