Fix loop structure and increase number of points in pass computation
The previous loop structure was very inefficient, as some things were needlessly repeated, such as checking whether a given point is valid, which was repeated for each new robot being considered as a target for the pass. Obviously, this does not affect the validity of a point in the field, and therefore should be outside the loop.

Furthermore, now that the scoring is more efficient, we can afford to increase the number of points considered. When using a grid of 81 points, this functions takes approximately 1 to 2 ms, which should be sufficiently quick. Perhaps this number can be increased further, or the spread of the points could still be improved (i.e. not being evenly spread but more dense in their half, etc.).