Create test fixture to ensure default environment for tests
This test fixture can be used while writing tests, to make sure that variables such as settings, and gamestates are always default when executing a new test. This solves the problem of not knowing what the state of the program is when writing is, as other tests could influence each other. This test fixture class is also a convenient place to create shared utility functions used by many tests. These can be accessed within the test as if the test was writting in this RTT_AI_tests class.