Refactor passComputations to be more readable and to return more information in a struct
In order to properly calculate a pass, we need to know which robot are the keeper and passer. However, these roles are not yet devided at the time of computing the pass. Hence, this new pass computation calculates which robots should have these roles and return their id's in the passInfo struct. These id's can then be set in the play as forced id's to make sure these are indeed set correctly.