Change PlayEvaluator to take a ptr to world as input instead of storing internally
Previously, a ptr to world was stored in the PlayEvaluator, which was then used in several places to get access to the world. We can avoid having to set/update the world field in the PlayEvaluator and also prevent having to get the world from the PlayEvaluator in several places by just passing along world as an argument. This also allows us to make the functions in this class static.