Set the world in each play at start, update field each tick
Since world is a static class, it is guaranteed that world will be at the same address for the lifetime of the program, it is therefore sufficient to set the pointer to world in each play once at the start of the program. The field, however, does need to be updated each tick, as this is not stored as a ptr and can change at any moment.