Pass field as argument for score() instead of playEvaluator
The only reason playEvaluator is passed as an argument, is to retrieve the world and field from it (and to call methods that are now static). The world, however, is by definition the same in both classes, since it is static. The field does differ, since the field in the play class is only updated when the play is being executed, but instead of getting it from the playEvaluator, it can be passed as an argument.