Do not use same margin for checking def area and field
The def area margin expands the area outwards, essentially making sure we are further away from the def area than we need to be as a safety measure. The field margin also expands the field, but this causes us to be allowed to be OUTSIDE of the field by that margin, which is not what we want. Thus, this margin should not be used the same for the field.

In the future, we should have a nicer system, where we can separately pass a field and def area margin to avoid this issue