Hotfix: don't apply defense area margin to our defense area
Whilst we can't be in the opposing defense area, we are allowed to be partly in our defense area and still get the ball (but not WHOLLY within). Thus, by not using a margin, we can allow ourselves to still get the ball even though it is right outside our defense area. This avoids the issue where neither the keeper nor any other robot can access the ball when it is just outside the defense area. In the future, this should be implemented in a nicer way perhaps, since this is obviously a giant hack.