Make keeper ballAvoider and remove positions to move to
It is easier to just have the robots stay where they are instead of moving them to these positions. Usually they are in a pretty decent position when the ball placement is called. Furthermore, the keeper should also be a ball avoider, as ballPlacement can go through the keeper area