Add extra margin to avoid point being between goal line and def area
Previously, it was possible to have a point (i.e. (-6.0, 1), that is within the field (as -6.0 >= -6.0), but NOT in the defense area (as for some reason, the defense area begins at -5.9 currently). This is bad since that point is then seen as a valid position, even though it is not. When the field class is refactored, which Floris has been working on, this issue should hopefully be fixed. But at least until then, this is a necessary extra safety.