Send kick command 5 times to ensure kick is executed
Just sending the kick command once might not be enough due to packet drops, delays, etc. and it was noticed that pressing the kick button quite often not lead to the robot kicking. By sending the command 5 times, it is more likely that the robot actually kicks after the button is pressed.