Fix widget locking world for longer than necessary
`rtt::world::World::instance()` returns a guard lock, which gets released
at the end of the scope (used to be once the paintEvent returned). This
causes STP to be delayed if `Input::getDrawings()` vector is large (over
3000). Since the lock release is postponed until the method finishes painting all data in the vector (despite not needing the world anymore).

Releasing the lock imminently after the field finished painting mitigates
this issue. As far as I can tell, that is the intended behavior since
`Input::getDrawings()` holds the separate lock. Thus, it does not require
the world to be locked during drawing.