Refactor ball to better handle invisible balls
If the ball is not visible, we now assume that the ball is at the same place it was previous tick. If a robot seems to have the ball, furthermore, we place the ball in front of that robot.

This allows us to deal with the scenario where a robot grabs a ball that we can not see, and then tries to rotate with that ball. Now, the ball will be moved together with the robot, ensuring that the robot doesn't think it lost the ball just because it can't see the ball moving.