Robot drives back without rotating after ballPlacement
Added new tactic to the role of BallPlacer call BallStandBack. This tactic is executed after robot places a ball in the designated spot. After doing so robot goes back without any rotation to the point that is control_constants::AVOID_BALL_DISTANCE away from the ball and then AvoidBall tactic is executed