Update hasBallDistance and max commanded vel
The hasBallDistance can be increased slightly, as a false positive is not as much of an issue with the new ball posession filtering. 0.11 seems like a good balance based on short tests, but this value might need to be tweaked further.

Finally, the max vel command is reduced to 4, which, for now, is a reasonable max velocity to send, as we never really want to (or can) go faster.